Move response codes to their own package
This allows us to use the examplebroker on the brokers package without
having issues with import cycles.
